+++ gen-documentation.js
@@ -357,6 +357,7 @@ const mainPages = [
   },
   {
     dropdown: 'Rules',
+    id: 'id-rules',
     pages: [
       { content: 'content-rulesets.njk',
         title: 'Rulesets',
@@ -404,6 +405,7 @@ const mainPages = [
   },
   {
     dropdown: 'About',
+    id: 'id-about',
     pages: [
       { content: 'content-about-history.njk',
         title: 'History',
@@ -448,8 +450,9 @@ function createNavigation(pages) {
              data-bs-toggle="dropdown"
             href="#"
             role="button"
-            aria-expanded="false">${item.dropdown}</a>
-          <ul class="dropdown-menu">`;
+            aria-expanded="false"
+            aria-controls="${item.id}">${item.dropdown}</a>
+          <ul class="dropdown-menu" id="${item.id}">`;
 
       item.pages.forEach( p => {
         console.log(`[dropdown][page]: ${p.filename}`);